changed costs from int to float

+++ main.c
@@ -50,9 +50,10 @@
 
 int main() {
     // Testing mealExpenses functions
-    int totalMealCost = 0;
-    int allowedMealCost = 0;
+    // TODO: Unit testing
+    float totalMealCost = 0;
+    float allowedMealCost = 0;
     calculateMealCosts(3,900,1200,&totalMealCost,&allowedMealCost);
-    printf("%d,%d",totalMealCost,allowedMealCost);
+    printf("%f,%f",totalMealCost,allowedMealCost);
     return 0;
 }
