+++ mealExpenses.h
@@ -1,10 +1,10 @@
 #ifndef MEALEXPENSES_H
 #define MEALEXPENSES_H
 
-void calculateMealCosts(int numDays, int departureTime, int arrivalTime, int *totalMealCost, int *allowableMealCost);
-int promptBreakfast(int day);
-int promptLunch(int day);
-int promptDinner(int day);
-int promptCost();
+void calculateMealCosts(int numDays, int departureTime, int arrivalTime, float *totalMealCost, float *allowableMealCost);
+float promptBreakfast(int day);
+float promptLunch(int day);
+float promptDinner(int day);
+float promptCost();
 
 #endif
